Ajout de la page library

+++ src/Controller/BooksController.php
@@ -19,8 +19,9 @@ class BooksController extends AbstractController
      */
     public function index(BookRepository $bookRepository)
     {
-        $books = $bookRepository->findAll();
-        return $this->render('books/index.html.twig', compact('books'));
+        $title = 'Derniers livres';
+        $books = $bookRepository->findBy([], ['createdAt' => 'DESC']);
+        return $this->render('books/index.html.twig', compact('title', 'books'));
     }
 
     /**
@@ -47,9 +48,17 @@ class BooksController extends AbstractController
     /**
      * @Route("/books/{id<[0-9]+>}", name="app_books_show", methods={"GET"})
      */
-    public function show(Book $book): Response
-    {
-        return $this->render('books/show.html.twig', compact('book'));
+    public function show(Book $book, Security $security): Response
+    {   
+        $isInLibrary = false;
+
+        if ($security->getUser()) {
+            if ($security->getUser()->getBooks()->contains($book)) {
+                $isInLibrary = true;
+            }
+        }
+
+        return $this->render('books/show.html.twig', compact('book', 'isInLibrary'));
     }
 
     /**
@@ -81,8 +90,12 @@ class BooksController extends AbstractController
     /**
      * @Route("/books/{id<[0-9]+>}", name="app_books_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, Book $book, EntityManagerInterface $em): Response
+    public function delete(Request $request, Book $book, EntityManagerInterface $em, Security $security): Response
     {
+        if ($security->getUser() !== $book->getCreatedBy()) {
+            return $this->redirectToRoute('app_home');
+        }
+
         if ($this->isCsrfTokenValid('book_deletion_'.$book->getId(), $request->request->get('csrf_token'))) {
             $em->remove($book);
             $em->flush();
@@ -92,5 +105,43 @@ class BooksController extends AbstractController
        
 
        return $this->redirectToRoute('app_home');
+    }
+
+    /**
+     * @Route("/books/{id<[0-9]+>}/add", name="app_books_add", methods={"PUT"})
+     */
+    public function add(Book $book, Request $request, EntityManagerInterface $em, Security $security): Response
+    {
+        if (!$security->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        if ($this->isCsrfTokenValid('book_add_'.$book->getId(), $request->request->get('csrf_token'))) {
+            $security->getUser()->addBook($book);
+            $em->flush();
+
+            $this->addFlash('info', 'Livre ajouté');
+        }
+
+        return $this->redirect($this->generateUrl('app_books_show', ['id' => $book->getId()]));
+    }
+
+    /**
+     * @Route("/books/{id<[0-9]+>}/remove", name="app_books_remove", methods={"DELETE"})
+     */
+    public function remove(Book $book, Request $request, EntityManagerInterface $em, Security $security): Response
+    {
+        if (!$security->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        if ($this->isCsrfTokenValid('book_remove_'.$book->getId(), $request->request->get('csrf_token'))) {
+            $security->getUser()->removeBook($book);
+            $em->flush();
+
+            $this->addFlash('info', 'Livre retiré');
+        }
+
+        return $this->redirect($this->generateUrl('app_books_show', ['id' => $book->getId()]));
     }
 }
